Added -ef flag, and -args flag, for extra args to ffmpeg. Video should probably be run through the ffmpeg side of the program for now.

+++ ContentConverter.py
@@ -49,6 +49,7 @@ params={}
 examples=[]
 stuff=sys.argv
 #Yes, I did write my own flag handler, and yes, I am a genuine idiot. But I don't care. This also allows me to set up defaults and stuff in the way I like.
+#Look at the current calls for examples. Note that maxargs=-1 reads up to the next -flag structure in argv.
 def grabFlag(args,flag,fields,defaults,maxargs):
 	if(flag in args):
 		values=[]
@@ -95,17 +96,21 @@ stuff=grabFlag(stuff,"-a",['type','extension','outextension'],['audio','.wav','.
 
 if('type' not in params):
 	stuff=grabFlag(stuff,"-standard",['max_size','outextension','rescale'],[100000,'.jpg',True],0) #Grab the -standard flag
-
 elif (params['type']=='video'):
-	stuff=grabFlag(stuff,"-standard",['max_size','outextension','warningfile'],[100000000,'.m4v','toobig.out'],0)
+	stuff=grabFlag(stuff,"-standard",['max_size','outextension','warningfile'],[100000000,'.m4v','toobig.out'],0) #For video
 elif (params['type']=='audio'):
-	stuff=grabFlag(stuff,"-standard",['max_size','outextension','warningfile'],[10,'.mp3','toobig.out'],0)
+	stuff=grabFlag(stuff,"-standard",['max_size','outextension','warningfile'],[10,'.mp3','toobig.out'],0) #And for audio
 
 stuff=grabFlag(stuff,"-s",['max_size'],[None],1) #Grab the -s flag
 stuff=grabFlag(stuff,"-i",['extension'],[None],1) #Grab the -i flag
 stuff=grabFlag(stuff,"-o",['outextension'],[None],1) #Grab the -o flag
 stuff=grabFlag(stuff,"-sw",['warningfile'],[None],1) #Grab the -sw flag
 stuff=grabFlag(stuff,"-l",['logfile'],[None],1) #Grab the -l flag
+stuff=grabFlag(stuff,"-ef",['errorfile'],[None],1) #Grab the -ef flag
+stuff=grabFlag(stuff,"-args",['extra_args'],[None],1) #Grab extra args for ffmpeg??
+if('extra_args' not in params):
+	params['extra_args']=''
+print stuff
 #Display help message
 if("-h" in stuff):
 	help=open("help.txt","r")
@@ -124,23 +129,29 @@ top=sys.argv[1]
 params['top']=top
 print top
 print params
-if ('max_size' in params):
+if ('max_size' in params): #Make sure that the size is an integer
 	try:
 		params['max_size']=int(params['max_size'])
 	except:
 		print "Size given was not an int"
 		sys.exit()
-if ('warningfile' in params):
+if ('warningfile' in params): #And that the oversize log file exists
 	try:
 		params['warningfile']=open(params['warningfile'],"w")
 	except:
 		print "Invalid file listed for size warnings"
 		sys.exit()
-if('logfile' in params):
+if('logfile' in params): #And the normal log file too.
 	try:
 		params['logfile']=open(params['logfile'],"w")
 	except:
 		print "Invalid log file"
+		sys.exit()
+if('errorfile' in params): #And the file for errors.
+	try:
+		params['errorfile']=open(params['errorfile'],"w")
+	except:
+		print "Invalid error file"
 		sys.exit()
 #If they are bad at typing directories, don't bother doing anything.
 try:
@@ -149,6 +160,7 @@ except:
 	logOutput("Directory "+str(top)+" not found.",params)
 	sys.exit()
 #Otherwise, walk through directory tree, and add .tiffs to examples
+#This section preserves some old functionality. For images, it was trained to look at the header using imghdr, in addition to the extension. Probably should be removed or changed to a flag.
 for root,dirs,files in os.walk(top):
 	for test in files:
 		if('extension' not in params):
@@ -162,6 +174,7 @@ for root,dirs,files in os.walk(top):
 		elif (test.endswith(params['extension']) and (root+"/"+test).find('/data/meta')==-1):
 			logOutput("Found "+root+"/"+test+" with correct type",params)
 			examples.append(root+"/"+test)
+#Pick the conversion tool based on type.
 if('type' in params):
 	if('type'=='video'):
 		HB.convertVideo(examples,params)
